Top refilling rework
Target is using
TopUpdater#reloadTops_UsingStatFiles()
instead of
TopUpdater#reloadTops_OfflinePlayerLoading().
Also removed russian-style decorators(would be replaced with formatted strings)

+++ src/com/te/top/RatingPlace.java
@@ -4,10 +4,10 @@ import java.util.ArrayList;
 import java.util.List;
 
 public class RatingPlace {
-	int place_value = 0;
+	long place_value = 0;
 	List<String> nicknames = new ArrayList<>();
 	
-	public RatingPlace(int value, String nick) {
+	public RatingPlace(long value, String nick) {
 		place_value = value;
 		nicknames.add(nick);
 	}
